requirements.txt
tkinterdnd2
GitPython
